+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Update API to use structs
 - Print branch name as well as URL when creating issues
+- Add `priority` flag to `issue create` and `template evaluate`
 
 ## 2023-11-20 v0.1.9